feat: Add 'Why Us?' blog post with navigation
- Install React Router DOM for client-side routing
- Create compelling WhyUs component with detailed content about choosing Elevated Deck Co
- Add comprehensive sections covering licensing, experience, quality materials, customer service, speed, and warranty
- Include ROI information showing 85% return on investment and 5K+ home value increase
- Add customer testimonials and guarantees
- Integrate 'Why Us?' navigation link in header with responsive design
- Set up routing configuration in App.tsx with Home and WhyUs routes
- Create separate Home component from existing App content
- Add .gitignore file for proper version control
- Implement seamless navigation between pages with consistent branding
- Include call-to-action sections and contact information
- All components follow existing design system and color scheme
